+++ examples/tanstack-example/vite.config.ts
@@ -1,12 +1,12 @@
 import { tanstackStart } from "@tanstack/react-start/plugin/vite";
-import { defineConfig } from "vite";
+import { defineConfig, esmExternalRequirePlugin } from "vite";
 import viteReact from "@vitejs/plugin-react";
 import tailwindcss from "@tailwindcss/vite";
 import { nitro } from "nitro/vite";
 
 export default defineConfig({
   optimizeDeps: {
-    include: ["tsdkarc", "tsdkarc-x"],
+    include: ["tsdkarc", "tsdkarc-x", "typescript"],
   },
   server: {
     port: 3000,
@@ -16,6 +16,9 @@ export default defineConfig({
   },
   plugins: [
     tailwindcss(),
+    esmExternalRequirePlugin({
+      external: ["typescript"],
+    }),
     tanstackStart({
       srcDirectory: "src",
     }),